pyproject.toml
[project]
name = "multi-agent-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.131.0",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
